+++ CHANGELOG.md
@@ -4,13 +4,18 @@ All notable changes to the "watermelon" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [1.6.0]
+## [1.6.1]
+- Get code context button now retrieves the most relevant Jira ticket, if the user is authenticated with Jira
+- Daily summary now shows currently open Jira tickets that are assigned to the user
+- New loading animation
+- Allows the user to deactivate telemetry through settings
+- Fixes stacking of elements in the extension
 
+## [1.6.0]
 - Watermelon is now an authentication provider, to allow us to connect to all your services
 - Made improvements to request less scripts from the internet, improving performance and security
 
 ## [1.5.3]
-
 - Will allow the user to use some features without logging into GitHub
 
 ## [1.5.2]